# src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: WEX Purchase Transaction Service
  version: 1.1.0
servers:
  - url: http://localhost:8080

paths:
  /api/v1/purchases:
    post:
      tags:
        - Purchases
      summary: Store a purchase transaction
      operationId: createPurchase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PurchaseCreateRequest'
      responses:
        '201':
          description: Purchase created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/purchases/{id}:
    get:
      tags:
        - Purchases
      summary: Retrieve a purchase converted using Treasury rates for a specific country-currency descriptor
      operationId: getPurchaseConverted
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid

        # No mapping: caller passes Treasury "country_currency_desc" exactly as downloaded at startup.
        - name: country_currency_desc
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 200

      responses:
        '200':
          description: Converted purchase
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseConvertedResponse'
        '400':
          description: Validation or conversion error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Purchase not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Treasury reference data unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    PurchaseCreateRequest:
      type: object
      required: [description, transactionDate, amountUsd]
      properties:
        description:
          type: string
          maxLength: 50
        transactionDate:
          type: string
          format: date
        amountUsd:
          type: string
          example: "100.10"


    PurchaseResponse:
      type: object
      required: [id, description, transactionDate, amountUsd]
      properties:
        id:
          type: string
          format: uuid
        description:
          type: string
        transactionDate:
          type: string
          format: date
        amountUsd:
          type: string

    PurchaseConvertedResponse:
      type: object
      required: [id, description, transactionDate, amountUsd, targetCurrency, exchangeRateUsed, convertedAmount]
      properties:
        id:
          type: string
          format: uuid
        description:
          type: string
        transactionDate:
          type: string
          format: date
        amountUsd:
          type: string

        # Store the descriptor here (no mapping).
        targetCurrency:
          type: string

        exchangeRateUsed:
          type: number
          format: double

        convertedAmount:
          type: number
          format: double

    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Machine-readable error category
        message:
          type: string
          description: Human-readable message